[IMPROVE] change the type data of price and qty from Float to Integer

+++ src/main/java/com/enigma/wmb_api/dto/response/BillDetailResponse.java
@@ -10,6 +10,6 @@ import lombok.*;
 public class BillDetailResponse {
     private String id;
     private String menuId;
-    private Float qty;
-    private Float price;
+    private Integer qty;
+    private Integer price;
 }
